pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sable"
version = "0.1.0"
description = "Sable content production toolkit for crypto Twitter"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "anthropic>=0.25",
    "rich>=13.0",
    "httpx>=0.27",
    "Pillow>=10.0",
    "faster-whisper>=1.0",
    "python-frontmatter>=1.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
face = [
    "replicate>=0.25",
    "face-recognition>=1.3",
    "imagehash>=4.3",
]
pulse = [
    "python-Levenshtein>=0.23",
]
score = [
    "librosa>=0.10",
    "opencv-python>=4.9",
    "numpy>=1.26",
]
thumbnail = [
    "playwright>=1.40",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.9",
]
all = [
    "replicate>=0.25",
    "imagehash>=4.3",
    "python-Levenshtein>=0.23",
]

[project.scripts]
sable = "sable.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sable*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",   # line too long — not enforced yet
    "E402",   # module-level import not at top — CLI pattern uses deferred imports
    "F401",   # unused imports — some re-exported intentionally
]
exclude = ["tests/"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
check_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
exclude = ["tests/"]
